requirements.txt
streamlit
pandas
numpy
altair
tensorflow-cpu
h5py